+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 1.12.8
 ### 13-Sep-2023
 - Added a check for FileVault being enabled during Setup Assistant (for macOS 14 Sonoma) ([Pull Request No. 96](https://github.com/dan-snelson/Setup-Your-Mac/pull/96); thanks, Obi-@drtaru!)
+- Added `-L` to `curl` command when caching banner images (thanks for the suggestion, @bartreardon!)
+- Added `swiftDialogMinimumRequiredVersion` variable to more easily track the minimum build. ([Pull Request No. 98](https://github.com/dan-snelson/Setup-Your-Mac/pull/98); thanks, @GadgetGeekNI!)
+
 
 ## 1.12.7
 ### 09-Sep-2023